Separate injecting context data into instance to `collect_instance_data.py` to simplify Collectors.
- Change `familyId` to `family`
- Refactor implementation for Integrator from `plugin.py` to `integrate_asset.py` plug-in.
