+++ pyblish_magenta/plugin.py
@@ -16,52 +16,3 @@ class Extractor(pyblish.api.Extractor):
             instance.set_data('extractDir', value=extract_dir)
 
         return extract_dir
-
-
-class Integrator(pyblish.api.Integrator):
-
-    def compute_integrate_dir(self, instance):
-        # Get instance data
-        data = {'root': instance.data('root'),
-                'container': instance.data('container', None),
-                'asset': instance.data('asset')}
-
-        family_id = instance.data('familyId')
-        if not family_id:
-            raise pyblish.api.ConformError(
-                "No family id found on instance. Can't resolve asset path.")
-
-        # Get asset directory
-        schema = pyblish_magenta.schema.load()
-        output_template = "{0}.asset".format(family_id)
-        dir_path = schema.get(output_template).format(data)
-
-        return dir_path
-
-    def integrate(self, instance):
-
-        # Get the extracted directory
-        extract_dir = instance.data('extractDir')
-        if not extract_dir:
-            raise pyblish.api.ConformError("Cannot integrate if no `extractDir` "
-                                           "temporary directory found.")
-
-        # Define the integrate directory
-        integrate_dir = self.compute_integrate_dir(instance)
-        instance.set_data('integrateDir', value=integrate_dir)
-
-        # Copy the files/directories from extract directory to the integrate directory
-        self.log.info("Integrating extracted files for '{0}'..".format(instance))
-        if os.path.isdir(integrate_dir):
-            self.log.info("Existing directory found, merging..")
-            for fname in os.listdir(extract_dir):
-                abspath = os.path.join(extract_dir, fname)
-                commit_path = os.path.join(integrate_dir, fname)
-                shutil.copy(abspath, commit_path)
-        else:
-            self.log.info("No existing directory found, creating..")
-            shutil.copytree(extract_dir, integrate_dir)
-
-        self.log.info("Integrated to directory '{0}'".format(integrate_dir))
-
-        return integrate_dir
