+++ pyblish_magenta/plugins/collect_review.py
@@ -27,14 +27,6 @@ class CollectCameras(pyblish.api.Collector):
             instance = context.create_instance(name=name, family="review")
             instance.add(camera)
 
-            # Set pipeline data
-            #data = schema.get("model.dev").parse(scene_name)
-            #instance.set_data("root", data["root"])
-            #instance.set_data("source_file", scene_name)
-            #instance.set_data("asset", data["asset"])
-            #instance.set_data("container", data["container"])
-            instance.set_data("familyId", context.data('familyId'))
-
             self.log.info("Found: {0}".format(camera))
 
             attrs = cmds.listAttr(camera, userDefined=True) or list()
